review comments and vault client updates
