+++ agent/connect/ca/provider_vault_auth.go
@@ -3,15 +3,8 @@ package ca
 import (
 	"context"
 	"fmt"
-	"os"
-	"strconv"
 
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/hashicorp/go-cleanhttp"
-	"github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/go-secure-stdlib/awsutil"
 	"github.com/hashicorp/vault/api"
-	"github.com/hashicorp/vault/api/auth/gcp"
 
 	"github.com/hashicorp/consul/agent/structs"
 )
@@ -23,10 +16,7 @@ type VaultAuthenticator interface {
 }
 
 // LoginDataGenerator is used to generate the login data for a Vault login API request.
-type LoginDataGenerator interface {
-	// GenerateLoginData creates the login data from the given authMethod configuration.
-	GenerateLoginData(authMethod *structs.VaultAuthMethod) (map[string]interface{}, error)
-}
+type LoginDataGeneratorFn func(authMethod *structs.VaultAuthMethod) (map[string]any, error)
 
 var _ VaultAuthenticator = (*VaultAuthClient)(nil)
 
@@ -39,18 +29,14 @@ type VaultAuthClient struct {
 	LoginPath string
 	// LoginDataGen derives the parameter map containing the data for the login API request.
 	// For some auth methods this is needed to transform the AuthMethod.Params into the login data
-	// needed for the API request.
-	LoginDataGen LoginDataGenerator
+	// used for the API request.
+	LoginDataGen LoginDataGeneratorFn
 }
 
 // NewVaultAPIAuthClient creates a VaultAuthClient that uses the Vault API to perform a login.
 func NewVaultAPIAuthClient(authMethod *structs.VaultAuthMethod, loginPath string) *VaultAuthClient {
 	if loginPath == "" {
-		mount := authMethod.Type
-		if authMethod.MountPath != "" {
-			mount = authMethod.MountPath
-		}
-		loginPath = fmt.Sprintf("auth/%s/login", mount)
+		loginPath = fmt.Sprintf("auth/%s/login", authMethod.MountPath)
 	}
 	return &VaultAuthClient{
 		AuthMethod: authMethod,
@@ -66,7 +52,7 @@ func (c *VaultAuthClient) Login(ctx context.Context, client *api.Client) (*api.S
 	// If a login data generator is provided then use it to transform the auth method params
 	// into the appropriate login data for the API request.
 	if c.LoginDataGen != nil {
-		loginData, err = c.LoginDataGen.GenerateLoginData(c.AuthMethod)
+		loginData, err = c.LoginDataGen(c.AuthMethod)
 		if err != nil {
 			return nil, fmt.Errorf("aws auth failed to generate login data: %w", err)
 		}
@@ -76,141 +62,6 @@ func (c *VaultAuthClient) Login(ctx context.Context, client *api.Client) (*api.S
 		ctx = context.Background()
 	}
 	return client.Logical().WriteWithContext(ctx, c.LoginPath, loginData)
-}
-
-var _ VaultAuthenticator = (*gcp.GCPAuth)(nil)
-
-// NewGCPAuthClient returns a VaultAuthenticator that can log into Vault using the GCP auth method.
-func NewGCPAuthClient(authMethod *structs.VaultAuthMethod) (VaultAuthenticator, error) {
-	// Check if the configuration already contains a JWT auth token. If so we want to
-	// perform a direct request to the login API with the config that is provided.
-	// This supports the  Vault CA config in a backwards compatible way so that we don't
-	// break existing configurations.
-	if containsVaultLoginParams(authMethod, "jwt") {
-		return NewVaultAPIAuthClient(authMethod, ""), nil
-	}
-
-	// Create a GCPAuth client from the CA config
-	params, err := toMapStringString(authMethod.Params)
-	if err != nil {
-		return nil, fmt.Errorf("misconfiguration of GCP auth parameters: %w", err)
-	}
-
-	opts := []gcp.LoginOption{gcp.WithMountPath(authMethod.MountPath)}
-
-	// Handle the login type explicitly.
-	switch params["type"] {
-	case "gce":
-		opts = append(opts, gcp.WithGCEAuth())
-	default:
-		opts = append(opts, gcp.WithIAMAuth(params["service_account_email"]))
-	}
-
-	auth, err := gcp.NewGCPAuth(params["role"], opts...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create a new Vault GCP auth client: %w", err)
-	}
-	return auth, nil
-}
-
-// NewAWSAuthClient returns a VaultAuthClient that can log into Vault using the AWS auth method.
-func NewAWSAuthClient(authMethod *structs.VaultAuthMethod) *VaultAuthClient {
-	authClient := NewVaultAPIAuthClient(authMethod, "")
-
-	// Inspect the config params to see if they are already in the format required for
-	// the auth/aws/login API call. If so, we want to call the login API with the params
-	// exactly as they are configured. This supports the  Vault CA config in a backwards
-	// compatible way so that we don't break existing configurations.
-	keys := []string{
-		"identity",                // EC2 identity
-		"pkcs7",                   // EC2 PKCS7
-		"iam_http_request_method", // IAM
-	}
-	if containsVaultLoginParams(authMethod, keys...) {
-		return authClient
-	}
-
-	// None of the known config is present so use the login data generator to transform
-	// the auth method config params into the login data request needed to authenticate
-	// via AWS.
-	authClient.LoginDataGen = &AWSLoginDataGenerator{}
-	return authClient
-}
-
-// AWSLoginDataGenerator is a LoginDataGenerator for AWS authentication.
-type AWSLoginDataGenerator struct {
-	// credentials supports explicitly setting the AWS credentials to use for authenticating with AWS.
-	// It is un-exported and intended for testing purposes only.
-	credentials *credentials.Credentials
-}
-
-// GenerateLoginData derives the login data for the Vault AWS auth login request from the CA
-// provider auth method config.
-func (g *AWSLoginDataGenerator) GenerateLoginData(authMethod *structs.VaultAuthMethod) (map[string]interface{}, error) {
-	// Create the credential configuration from the parameters.
-	credsConfig, headerValue, err := NewCredentialsConfig(authMethod.Params)
-	if err != nil {
-		return nil, fmt.Errorf("aws auth failed to create credential configuration: %w", err)
-	}
-
-	// Use the provided credentials if they are present.
-	creds := g.credentials
-	if creds == nil {
-		// Generate the credential chain
-		creds, err = credsConfig.GenerateCredentialChain()
-		if err != nil {
-			return nil, fmt.Errorf("aws auth failed to generate credential chain: %w", err)
-		}
-	}
-
-	// Derive the login data
-	loginData, err := awsutil.GenerateLoginData(creds, headerValue, credsConfig.Region, hclog.NewNullLogger())
-	if err != nil {
-		return nil, fmt.Errorf("aws auth failed to generate login data: %w", err)
-	}
-	return loginData, nil
-}
-
-// NewCredentialsConfig creates an awsutil.CredentialsConfig from the given set of parameters.
-// It is mostly copied from the awsutil package because that package does not currently provide
-// an easy way to configure all the available options.
-func NewCredentialsConfig(config map[string]interface{}) (*awsutil.CredentialsConfig, string, error) {
-	params, err := toMapStringString(config)
-	if err != nil {
-		return nil, "", fmt.Errorf("misconfiguration of AWS auth parameters: %w", err)
-	}
-
-	// Populate the AWS credentials configuration. Empty strings will use the defaults.
-	c := &awsutil.CredentialsConfig{
-		AccessKey:            params["access_key"],
-		SecretKey:            params["secret_key"],
-		SessionToken:         params["session_token"],
-		IAMEndpoint:          params["iam_endpoint"],
-		STSEndpoint:          params["sts_endpoint"],
-		Region:               params["region"],
-		Filename:             params["filename"],
-		Profile:              params["profile"],
-		RoleARN:              params["role_arn"],
-		RoleSessionName:      params["role_session_name"],
-		WebIdentityTokenFile: params["web_identity_token_file"],
-		HTTPClient:           cleanhttp.DefaultClient(),
-	}
-
-	if maxRetries, err := strconv.Atoi(params["max_retries"]); err == nil {
-		c.MaxRetries = &maxRetries
-	}
-
-	if c.Region == "" {
-		c.Region = os.Getenv("AWS_REGION")
-		if c.Region == "" {
-			c.Region = os.Getenv("AWS_DEFAULT_REGION")
-			if c.Region == "" {
-				c.Region = "us-east-1"
-			}
-		}
-	}
-
-	return c, params["header_value"], nil
 }
 
 func toMapStringString(in map[string]interface{}) (map[string]string, error) {
